requirements.txt
tensorflow==2.12.0
streamlit==1.10.0
Pillow==9.5.0
